modified setting and directory structure
