+++ package.json
@@ -18,6 +18,7 @@
         "node-sass": "^4.13.1",
         "react": "^16.13.0",
         "react-dom": "^16.13.0",
+        "react-router-dom": "^6.0.0-beta.0",
         "react-scripts": "3.4.0",
         "typescript": "^4.0.2"
     },
@@ -45,10 +46,17 @@
         ]
     },
     "devDependencies": {
+        "@types/eslint": "^6.8.1",
+        "@types/gh-pages": "^2.2.0",
+        "@types/material-ui": "^0.20.8",
+        "@types/node-sass": "^4.11.1",
+        "@types/testing-library__jest-dom": "^5.9.5",
+        "@types/testing-library__react": "^10.2.0",
         "@typescript-eslint/eslint-plugin": "^2.26.0",
         "@typescript-eslint/parser": "^2.26.0",
         "eslint": "^6.8.0",
         "eslint-plugin-react": "^7.19.0",
-        "gh-pages": "^2.2.0"
+        "gh-pages": "^2.2.0",
+        "typesync": "^0.8.0"
     }
 }
