+++ src/components/common/UsefulHooks.ts
@@ -1,0 +1,23 @@
+import { useState, useEffect } from 'react';
+
+export const useWindowDimensions: any = () => {
+  const getWindowDimensions = () => {
+    const { innerWidth: width, innerHeight: height } = window;
+    return {
+      width,
+      height,
+    };
+  };
+
+  const [windowDimensions, setWindowDimensions] = useState(
+    getWindowDimensions(),
+  );
+  useEffect(() => {
+    const onResize = () => {
+      setWindowDimensions(getWindowDimensions());
+    };
+    window.addEventListener('resize', onResize);
+    return () => window.removeEventListener('resize', onResize);
+  }, []);
+  return windowDimensions;
+};
