+++ src/components/styles/_colors.scss
@@ -1,5 +1,6 @@
 $polygon-pink: #ee5a62;
 $polygon-blue: #4a9cd5;
-$background-grey: rgba(0, 0, 0, .7);
-$hover-white: rgba(255, 255, 255, .3);
-
+$background-grey: rgba(0, 0, 0, 0.7);
+$hover-white: rgba(255, 255, 255, 0.3);
+$favicon-blue: rgba(0, 110, 255, 0.8);
+$hover-favicon-blue: rgba(80, 156, 255, 0.3);
