+++ src/components/styles/mixin.scss
@@ -1,0 +1,40 @@
+@import 'colors';
+
+.container {
+  width: 100%;
+  height: 100%;
+  padding-right: 1rem;
+  background-color: $background-grey;
+}
+
+.content-wrapper {
+  padding: 5rem 0;
+  color: white;
+  font-size: 1.2rem;
+}
+
+.title {
+  font-size: 4rem;
+  font-family: Anton, sans-serif;
+  text-align: center;
+  padding: 1rem, 0;
+}
+
+.table-wrapper {
+  display: grid;
+  flex-wrap: wrap;
+  justify-content: center;
+  border-spacing: 0.5rem;
+  border-collapse: separate;
+
+  tr {
+    td {
+      padding: 8px;
+    }
+    li {
+      font-size: 1.2rem;
+      padding-bottom: 0.5rem;
+      flex-wrap: nowrap;
+    }
+  }
+}
