+++ src/index.css
@@ -1,4 +1,4 @@
 body {
-    margin: 0 auto;
-    padding: 0;
+  margin: 0 auto;
+  padding: 0;
 }
